LayoutTests/editing/inserting/insert-paste-plain-text-into-list-item-expected.txt

This tests that pasting plain text into an empty list item produces flat sibling list items, not nested ones

Before:
| "\n"
| <ul>
|   <li>
|     <#selection-caret>
|     <br>
| "\n"

After:
| "\n"
| <ul>
|   <li>
|     "Line one"
|   <li>
|     "Line two"
|   <li>
|     "Line three"
|   <li>
|     <#selection-caret>
| "\n"
